requirements.txt
numba
numpy
nuscenes-devkit
opencv-python-headless
pandas
pytorch-lightning==1.6.0
scikit-image
scipy
setuptools==59.5.0
tensorboardX
# torch==1.9.0
# torchvision==0.10.0
inplace_abn